requirements.txt
flask
bcrypt
mysql-connector-python
openpyxl